requirements.txt
hyper==0.7.0
numpy>=1.10.0
Orange3>=3.3.5
requests
cachecontrol
lockfile